add a removed variable to mines to make sure the mine is completely removed and the reset task is cancelled

+++ src/us/creepermc/mines/managers/ConfirmRemoveGUI.java
@@ -46,6 +46,10 @@ public class ConfirmRemoveGUI extends XInvManager {
 		player.removeMetadata("islandmines_mine", getCore());
 		switch(menuItem.getId()) {
 			case "confirm":
+				if(!mine.canRemove()) {
+					getCore().sendMsg(player, "MINE_SAFETY");
+					break;
+				}
 				storageManager.deleteMine(mine, player, true);
 				break;
 			case "deny":
